+++ src/engine/podman.rs
@@ -111,8 +111,10 @@ impl EnginePodman {
     pub fn remove<S: AsRef<str>>(&self, name: S) -> Result<()> {
         let podman = self.podman_bin.to_string();
         let name = self.n(name);
-        run_fun!($podman kill $name)?;
-        run_fun!($podman rm -f $name)?;
+        if run_cmd!($podman container exists $name).is_ok() {
+            run_fun!($podman kill $name)?;
+            run_fun!($podman rm -f $name)?;
+        }
 
         if let Some(pod) = &self.pod {
             if run_cmd!($podman pod exists $pod).is_ok()
